Allow fitting procedure on a different cluster and make the fitting database accessible via the MongoDB  (#314)
* add some drafts

* pre-commit auto-fixes

* restructure and rely on flow instead

* restructure and rely on flow instead

* pre-commit auto-fixes

* more restructuring of completeworkflow

* pre-commit auto-fixes

* more restructuring of completeworkflow

* bring the workflow more in shape

* pre-commit auto-fixes

* fix the workflow stepwise

* pre-commit auto-fixes

* fix some more probles

* pre-commit auto-fixes

* hard code displacement to 0.01 in benchmark, fix other more issues in workflow

* pre-commit auto-fixes

* fix more logic problems

* make outputs nicer

* pre-commit auto-fixes

* fix a bunch of tests in auto phonons after the new changes

* fix more tests

* fix benchmark tests

* daza

* fix more tests

* fix pre database position for all workflows

* add more documentaion and fix number of jobs, addition of get_output

* pre-commit auto-fixes

* add more documentation

* fix list comprehension

* pre-commit auto-fixes

* mace stuff

* add to data

* add to data

* fix ranom seed for structures that are too similar

* pre-commit auto-fixes

* fix ranom seed for structures that are too similar

* pre-commit auto-fixes

* default random seed

* default random seed

* fix write benchmark generation beyond runs in jobflow

* pre-commit auto-fixes

* fix random seed and add strict tests

* fix lorbit problems

* fix test

* fix test

* fix rms computation

* pre-commit auto-fixes

* add one more test

* add documentation

* add hint on the default displacement

* add hint on the default displacement

* fix phonon data generation

* pre-commit auto-fixes

* clean up

* fix references

* add test for stratisfied split

* fix doc issues and data issues

* pre-commit auto-fixes

* fix doc issues and data issues

* pre-commit auto-fixes

* fix import issues

* pre-commit auto-fixes

* fix import issues

* pre-commit auto-fixes

* add test for pre-database

* deactivate filter* on test data

* pre-commit auto-fixes

* Add tests for test file sizes

* pre-commit auto-fixes

* remove definition

* fix output of do_

* pre-commit auto-fixes

* extend test to more generations and fix dft reference passing

* pre-commit auto-fixes

* fix doc

* fix doc

* fix doc

* fix doc

* pre-commit auto-fixes

* try increasing myst heading anchors

* add option to fit on a different cluster

* pre-commit auto-fixes

* try adding back _all__ block

* fix serialization problem

* pre-commit auto-fixes

* fix linting

* fix serialization problem

* fix glue xml, remove likely wrong database_dir

* pre-commit auto-fixes

* add try except block in rmse calculation

* pre-commit auto-fixes

* fix some more None issues

* pre-commit auto-fixes

* fix some more None issues

* fix some more None issues

* pre-commit auto-fixes

* slightly change default names

* fix test names

* fix test names

* pre-commit auto-fixes

* fix test names

* fix test names

* fix test names

* fix test names

* fix test names

* fix test names

* fix more errors

* fix some more remote issues

* pre-commit auto-fixes

* fix some more remote issues

* test iterative_phonon_flow.py

* test iterative_phonon_flow.py fix

* fix output issue

* add some docu

* fix format

* add another test

* fix confusing list outputs

* pre-commit auto-fixes

* fix doc bug

* pre-commit auto-fixes

* fix merge conflict

* remove debug prints

* pre-commit auto-fixes

* improve doc text

* improve doc text

* some improvements in the iterative unit test

* some improvements in the iterative unit test

* adjust doc to new DFT job names

* a few improvements of the if self.run_fits_on_different_cluster block

* small rearrangement

* fix unit tests

* fix unit tests

---------

+++ src/autoplex/fitting/common/utils.py
@@ -1893,17 +1893,20 @@ def prepare_fit_environment(
     os.makedirs(
         os.path.join(mlip_path, train_name.replace("train.extxyz", "")), exist_ok=True
     )
-    shutil.copy(
-        os.path.join(database_dir, test_name),
-        os.path.join(mlip_path, test_name),
-    )
-    shutil.copy(
-        os.path.join(database_dir, train_name),
-        os.path.join(mlip_path, train_name),
-    )
-    if glue_xml:
+    if not Path(mlip_path / test_name).exists():
         shutil.copy(
-            os.path.join(database_dir, glue_name),
+            os.path.join(database_dir, test_name),
+            os.path.join(mlip_path, test_name),
+        )
+    if not Path(mlip_path / train_name).exists():
+        shutil.copy(
+            os.path.join(database_dir, train_name),
+            os.path.join(mlip_path, train_name),
+        )
+    if glue_xml:
+        # TODO: might need to be fixed for remote connection
+        shutil.copy(
+            Path(glue_name),
             os.path.join(mlip_path, "glue.xml"),
         )
